Add Authorize and filter

+++ app/Http/Controllers/api/DepartmentController.php
@@ -2,17 +2,22 @@
 
 namespace App\Http\Controllers\Api;
 
+use App\Filters\DepartmentFilter;
 use App\Http\Controllers\Api\BaseApiController;
 use App\Http\Requests\StoreDepartmentRequest;
 use App\Http\Requests\UpdateDepartmentRequest;
 use App\Http\Resources\DepartmentResource;
 use App\Models\Department;
+use Illuminate\Http\Request;
 
 class DepartmentController extends BaseApiController
 {
-    public function index()
+
+    public function index(Request $request)
     {
-        $departments = Department::with('images')->paginate(20);
+        $filters = new DepartmentFilter($request);
+        $departments = Department::with('images')->filter($filters)
+            ->paginate(20);;
         return $this->successResponse('Departments retrieved successfully', DepartmentResource::collection($departments));
     }
     public function store(StoreDepartmentRequest $request)
@@ -30,6 +35,7 @@ class DepartmentController extends BaseApiController
     }
     public function update(UpdateDepartmentRequest $request, Department $department)
     {
+        $this->authorize('update', $department);
         $data = $request->validated();
         $department->update($data);
         // i still didnt do the logic of image uploading
@@ -38,6 +44,7 @@ class DepartmentController extends BaseApiController
     }
     public function destroy(Department $department)
     {
+        $this->authorize('delete', $department);
         $department->delete();
         return $this->successResponse('Department deleted successfully');
     }
